添加人脸信息的定时扫描入库

+++ src/main/java/cn/edu/swpu/face_detection_register/model/vo/DetectFacePartVo.java
@@ -32,7 +32,7 @@ public class DetectFacePartVo implements Serializable {
 
     @ApiModelProperty(value = "美丑得分")
     @JsonProperty(value = "beauty")
-    private Double beauty;
+    private Integer beauty;
 
     @ApiModelProperty(value = "表情")
     @JsonProperty(value = "expression")
